Rather than have a dedicated ?roadtrip? table, added a ?collection? table with an option to specify different types: roadtrip, walking tour, audio guide.
Added a ?topic? table to the model.

Added a ?region? (or area? or siimply polygon?) column to the story table.

Added a date column to the story table.

Added an optional image field.

Added an optional audio field.

Removed the ?comment? field for stories in favor of a ?description? field.